Centralize web wasm loading
