+++ web/packages/accounts-rs/services/crypto.ts
@@ -1,4 +1,4 @@
-import { ensureCryptoInit, enteWasm } from "./wasm";
+import { ensureWasmCryptoInit, loadEnteWasm } from "ente-wasm/load";
 
 export const deriveKeyInsufficientMemoryErrorMessage =
     "Failed to derive key (insufficient memory)";
@@ -84,8 +84,8 @@ export const deriveKey = async (
     opsLimit: number,
     memLimit: number,
 ) => {
-    await ensureCryptoInit();
-    const wasm = await enteWasm();
+    await ensureWasmCryptoInit();
+    const wasm = await loadEnteWasm();
     return wasm.auth_derive_kek(password, saltB64, memLimit, opsLimit);
 };
 
@@ -99,9 +99,9 @@ const normalizeDerivedKey = (result: WasmDerivedKey): DerivedKey => ({
 export const deriveSensitiveKey = async (
     password: string,
 ): Promise<DerivedKey> => {
-    await ensureCryptoInit();
+    await ensureWasmCryptoInit();
     try {
-        const wasm = await enteWasm();
+        const wasm = await loadEnteWasm();
         return normalizeDerivedKey(
             wasm.auth_generate_sensitive_kek(password) as WasmDerivedKey,
         );
@@ -113,9 +113,9 @@ export const deriveSensitiveKey = async (
 export const deriveInteractiveKey = async (
     password: string,
 ): Promise<DerivedKey> => {
-    await ensureCryptoInit();
+    await ensureWasmCryptoInit();
     try {
-        const wasm = await enteWasm();
+        const wasm = await loadEnteWasm();
         return normalizeDerivedKey(
             wasm.auth_generate_interactive_kek(password) as WasmDerivedKey,
         );
@@ -125,14 +125,14 @@ export const deriveInteractiveKey = async (
 };
 
 export const generateKey = async (): Promise<string> => {
-    await ensureCryptoInit();
-    const wasm = await enteWasm();
+    await ensureWasmCryptoInit();
+    const wasm = await loadEnteWasm();
     return wasm.crypto_generate_key();
 };
 
 export const generateKeyPair = async (): Promise<KeyPair> => {
-    await ensureCryptoInit();
-    const wasm = await enteWasm();
+    await ensureWasmCryptoInit();
+    const wasm = await loadEnteWasm();
     const keyPair = wasm.crypto_generate_keypair();
     return { publicKey: keyPair.public_key, privateKey: keyPair.secret_key };
 };
@@ -141,8 +141,8 @@ export const encryptBox = async (
     dataB64: string,
     keyB64: string,
 ): Promise<EncryptedBox> => {
-    await ensureCryptoInit();
-    const wasm = await enteWasm();
+    await ensureWasmCryptoInit();
+    const wasm = await loadEnteWasm();
     const box = wasm.crypto_encrypt_box(dataB64, keyB64);
     return { encryptedData: box.encrypted_data, nonce: box.nonce };
 };
@@ -151,8 +151,8 @@ export const decryptBox = async (
     box: EncryptedBox,
     keyB64: string,
 ): Promise<string> => {
-    await ensureCryptoInit();
-    const wasm = await enteWasm();
+    await ensureWasmCryptoInit();
+    const wasm = await loadEnteWasm();
     return wasm.crypto_decrypt_box(box.encryptedData, box.nonce, keyB64);
 };
 
@@ -160,8 +160,8 @@ export const boxSealOpenBytes = async (
     encryptedData: string,
     keyPair: KeyPair,
 ): Promise<Uint8Array> => {
-    await ensureCryptoInit();
-    const wasm = await enteWasm();
+    await ensureWasmCryptoInit();
+    const wasm = await loadEnteWasm();
     return fromB64(
         wasm.crypto_box_seal_open(
             encryptedData,
@@ -177,8 +177,8 @@ export const deriveSubKeyBytes = async (
     subKeyID: number,
     context: string,
 ) => {
-    await ensureCryptoInit();
-    const wasm = await enteWasm();
+    await ensureWasmCryptoInit();
+    const wasm = await loadEnteWasm();
     return fromB64(
         wasm.crypto_derive_subkey(
             keyB64,
@@ -193,8 +193,8 @@ export const generateSRPSetupAttributesRust = async (
     kekB64: string,
     srpUserID: string,
 ) => {
-    await ensureCryptoInit();
-    const wasm = await enteWasm();
+    await ensureWasmCryptoInit();
+    const wasm = await loadEnteWasm();
     const setup = wasm.auth_generate_srp_setup(
         kekB64,
         srpUserID,
@@ -207,13 +207,13 @@ export const generateSRPSetupAttributesRust = async (
 };
 
 export const recoveryKeyFromMnemonicOrHex = async (value: string) => {
-    await ensureCryptoInit();
-    const wasm = await enteWasm();
+    await ensureWasmCryptoInit();
+    const wasm = await loadEnteWasm();
     return wasm.auth_recovery_key_from_mnemonic_or_hex(value);
 };
 
 export const recoveryKeyToMnemonicRust = async (recoveryKey: string) => {
-    await ensureCryptoInit();
-    const wasm = await enteWasm();
+    await ensureWasmCryptoInit();
+    const wasm = await loadEnteWasm();
     return wasm.auth_recovery_key_to_mnemonic(recoveryKey);
 };
